+++ src/com/studenttracker/model/Quiz.java
@@ -1,20 +1,19 @@
 package com.studenttracker.model;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class Quiz {
     private Integer quizId;
     private Integer lessonId;
     private byte[] quizPdfData;       // Binary PDF data (or String pdfPath if storing path)
-    private BigDecimal totalMarks;
+    private Double totalMarks;
     private LocalDateTime createdAt;
     private Integer createdBy;
     
     // Constructors
     public Quiz() {}
     
-    public Quiz(Integer lessonId, byte[] quizPdfData, BigDecimal totalMarks, Integer createdBy) {
+    public Quiz(Integer lessonId, byte[] quizPdfData, Double totalMarks, Integer createdBy) {
         this.lessonId = lessonId;
         this.quizPdfData = quizPdfData;
         this.totalMarks = totalMarks;
@@ -32,8 +31,8 @@ public class Quiz {
     public byte[] getQuizPdfData() { return quizPdfData; }
     public void setQuizPdfData(byte[] quizPdfData) { this.quizPdfData = quizPdfData; }
     
-    public BigDecimal getTotalMarks() { return totalMarks; }
-    public void setTotalMarks(BigDecimal totalMarks) { this.totalMarks = totalMarks; }
+    public Double getTotalMarks() { return totalMarks; }
+    public void setTotalMarks(Double totalMarks) { this.totalMarks = totalMarks; }
     
     public LocalDateTime getCreatedAt() { return createdAt; }
     public void setCreatedAt(LocalDateTime createdAt) { this.createdAt = createdAt; }
